requirements.txt
Flask==2.0.1
nltk==3.6.2
pytz==2021.1
tensorflow-cpu==2.5.0
gunicorn==20.1.0
